cv-server/requirements.txt
flask>=3.0.0
ultralytics>=8.3.0
pillow>=10.0.0
